requirements.txt
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
db-dtypes==1.2.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.47.2
google-api-core==2.16.2
google-auth==2.27.0
google-cloud-bigquery==3.17.1
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.18.1
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
nbformat==5.9.2
nest-asyncio==1.6.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pymsteams==0.2.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
rsa==4.9
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.6
wcwidth==0.2.13
zipp==3.17.0
